Update GameState.java JavaDocs
It's important for people to know what these GameStates actually mean, they need to be documented just like in spongepowered.api.event.Result.java. I don't know if these are exactly what the original author of this enum intended them to mean, but hopefully these can act as a point of discussion and base reference.